+++ e2e_tests/order/processing.ts
@@ -1,6 +1,4 @@
 import { ApiPromise, Keyring, WsProvider } from '@polkadot/api';
-import { KeyringPair } from '@polkadot/keyring/types';
-import { getEncodedRegionId, RegionId } from 'coretime-utils';
 import assert from 'node:assert';
 import {
   log,
@@ -13,8 +11,9 @@ import {
 } from '../common';
 import { UNIT } from '../consts';
 import { configureBroker, purchaseRegion, startSales } from '../coretime.common';
-import { ismpAddParachain, makeIsmpResponse, queryRequest } from '../ismp.common';
+import { ismpAddParachain } from '../ismp.common';
 import { REGIONX_API_TYPES, REGIONX_CUSTOM_RPC } from '../types';
+import { transferRegionToRegionX } from '../xc.common';
 
 async function run(_nodeName: string, networkInfo: any, _jsArgs: any) {
   const { wsUri: regionXUri } = networkInfo.nodesByName['regionx-collator01'];
@@ -52,17 +51,16 @@ async function run(_nodeName: string, networkInfo: any, _jsArgs: any) {
   // Needed for fee payment
   // Alice has relay tokens on Coretime chain by default, so no need to send there.
   log('Transfering rc token to RegionX:');
-  await transferRelayAssetToPara(UNIT, 2000, rococoApi, alice);
+  await transferRelayAssetToPara(rococoApi, alice, 1005, alice.address, 100n * UNIT);
+  await transferRelayAssetToPara(rococoApi, alice, 2000, alice.address, 100n * UNIT);
 
-  // 1. A region is listed on sale
+  // It takes time to cross-chain transfer assets.
+  await sleep(10000);
+
   log('Configuring coretime chain:');
   await configureBroker(coretimeApi, alice);
   log('Starting sales:');
   await startSales(coretimeApi, alice);
-
-  log('Giving alice balance on Coretime chain:');
-  const txSetBalance = coretimeApi.tx.balances.forceSetBalance(alice.address, 1000n * UNIT);
-  await submitExtrinsic(alice, coretimeApi.tx.sudo.sudo(txSetBalance), {});
 
   const regionId = await purchaseRegion(coretimeApi, alice);
   if (!regionId) throw new Error('RegionId not found');
@@ -70,7 +68,6 @@ async function run(_nodeName: string, networkInfo: any, _jsArgs: any) {
   log('Transferring region to RegionX');
   await transferRegionToRegionX(coretimeApi, regionXApi, alice, regionId);
 
-  // 2. An order is created
   const paraId = 2000;
   const orderRequirements = {
     begin: 40,
@@ -89,13 +86,11 @@ async function run(_nodeName: string, networkInfo: any, _jsArgs: any) {
     requirements: orderRequirements
   });
   
-  // 3. Fund the order
   log('Giving Bob tokens');
-  const giveBalanceCall = regionXApi.tx.tokens.setBalance(
+  const giveBalanceCall = regionXApi.tx.tokens.transfer(
     bob.address,
     RELAY_ASSET_ID,
-    200n * UNIT,
-    0
+    30n * UNIT,
   );
   await submitExtrinsic(alice, regionXApi.tx.sudo.sudo(giveBalanceCall), {});
 
@@ -104,92 +99,14 @@ async function run(_nodeName: string, networkInfo: any, _jsArgs: any) {
   const contributeCall = regionXApi.tx.orders.contribute(0, 10n * UNIT);
   await submitExtrinsic(bob, contributeCall, {});
 
-  // 4. A trader sees a profitable opportunity and fulfills the order.
   log('Alice fulfilling the order');
   const fulfillCall = regionXApi.tx.processor.fulfillOrder(0, regionId);
   await submitExtrinsic(alice, fulfillCall, {});
-  // TODO: check state
-
-  // 5. Ensure the region gets assigned to the specified parachain.
-}
-
-async function transferRegionToRegionX(
-  coretimeApi: ApiPromise,
-  regionXApi: ApiPromise,
-  sender: KeyringPair,
-  regionId: RegionId
-) {
-  const receiverKeypair = new Keyring();
-  receiverKeypair.addFromAddress(sender.address);
-
-  const feeAssetItem = 0;
-  const weightLimit = 'Unlimited';
-  const reserveTransferToRegionX = coretimeApi.tx.polkadotXcm.limitedReserveTransferAssets(
-    { V3: { parents: 1, interior: { X1: { Parachain: 2000 } } } }, //dest
-    {
-      V3: {
-        parents: 0,
-        interior: {
-          X1: {
-            AccountId32: {
-              chain: 'Any',
-              id: receiverKeypair.pairs[0].publicKey,
-            },
-          },
-        },
-      },
-    }, //beneficiary
-    {
-      V3: [
-        {
-          id: {
-            Concrete: {
-              parents: 1,
-              interior: 'Here',
-            },
-          },
-          fun: {
-            Fungible: 10n ** 10n,
-          },
-        }, // ^^ fee payment asset
-        {
-          id: {
-            Concrete: {
-              parents: 0,
-              interior: { X1: { PalletInstance: 50 } },
-            },
-          },
-          fun: {
-            NonFungible: {
-              Index: getEncodedRegionId(regionId, coretimeApi).toString(),
-            },
-          },
-        },
-      ],
-    }, //asset
-    feeAssetItem,
-    weightLimit
-  );
-  await submitExtrinsic(sender, reserveTransferToRegionX, {});
-
-  await sleep(5000);
-
-  const requestRecord = regionXApi.tx.regions.requestRegionRecord(regionId);
-  await submitExtrinsic(sender, requestRecord, {});
-
+  // Region should be removed after assigning it:
   let regions = await regionXApi.query.regions.regions.entries();
-  let region = regions[0][1].toHuman() as any;
-  assert(region.owner == sender.address);
-  assert(typeof region.record.Pending === 'string');
+  assert.equal(regions.length, 0);
 
-  // Respond to the ISMP get request:
-  const request = await queryRequest(regionXApi, region.record.Pending);
-  await makeIsmpResponse(regionXApi, coretimeApi, request, sender.address);
-
-  // The record should be set after ISMP response:
-  regions = await regionXApi.query.regions.regions.entries();
-  region = regions[0][1].toHuman() as any;
-  assert.equal(region.record.Available.paid, null);
+  // TODO: Ensure the region gets assigned to the specified parachain.
 }
 
 export { run };
