+++ e2e_tests/xc-transfer/asset-transfer.ts
@@ -51,7 +51,7 @@ async function run(nodeName: string, networkInfo: any, _jsArgs: any) {
   await assertRegionXBalance(alice.address, 0n);
   await assertRococoBalance(alice.address, 10n ** 18n);
 
-  await transferRelayAssetToPara(3n * 10n ** 12n, 2000, rococoApi, alice);
+  await transferRelayAssetToPara(rococoApi, alice, 2000, alice.address, 3n * 10n ** 12n);
   await sleep(5 * 1000);
 
   await assertRegionXBalance(alice.address, 3n * 10n ** 12n);
