+++ e2e_tests/xc-transfer/region-transfer.ts
@@ -1,19 +1,15 @@
 import { ApiPromise, Keyring, WsProvider } from '@polkadot/api';
-import { getEncodedRegionId } from 'coretime-utils';
-import assert from 'node:assert';
 import {
   openHrmpChannel,
   setupRelayAsset,
-  sleep,
   submitExtrinsic,
   transferRelayAssetToPara,
 } from '../common';
 import { UNIT } from '../consts';
 import { configureBroker, purchaseRegion, startSales } from '../coretime.common';
-import { ismpAddParachain, makeIsmpResponse, queryRequest } from '../ismp.common';
+import { ismpAddParachain } from '../ismp.common';
 import { REGIONX_API_TYPES, REGIONX_CUSTOM_RPC } from '../types';
-
-const REGIONX_SOVEREIGN_ACCOUNT = '5Eg2fntJ27qsari4FGrGhrMqKFDRnkNSR6UshkZYBGXmSuC8';
+import { transferRegionToCoretimeChain, transferRegionToRegionX } from '../xc.common';
 
 async function run(_nodeName: any, networkInfo: any, _jsArgs: any) {
   const { wsUri: regionXUri } = networkInfo.nodesByName['regionx-collator01'];
@@ -43,160 +39,22 @@ async function run(_nodeName: any, networkInfo: any, _jsArgs: any) {
   await openHrmpChannel(alice, rococoApi, 2000, 1005);
   await ismpAddParachain(alice, regionXApi);
 
-  // Needed for fee payment
-  // Alice has relay tokens on Coretime chain by default, so no need to send there.
-  await transferRelayAssetToPara(10n ** 12n, 2000, rococoApi, alice);
+  await transferRelayAssetToPara(rococoApi, alice, 1005, alice.address, 1000n * UNIT);
+  await transferRelayAssetToPara(rococoApi, alice, 2000, alice.address, 1000n * UNIT);
 
   await configureBroker(coretimeApi, alice);
   await startSales(coretimeApi, alice);
 
-  const txSetBalance = coretimeApi.tx.balances.forceSetBalance(alice.address, 1000n * UNIT);
-  await submitExtrinsic(alice, coretimeApi.tx.sudo.sudo(txSetBalance), {});
-
   const regionId = await purchaseRegion(coretimeApi, alice);
   if (!regionId) throw new Error('RegionId not found');
 
-  const receiverKeypair = new Keyring();
-  receiverKeypair.addFromAddress(alice.address);
+  // Transferring to the RegionX chain should work:
+  // NOTE: the function contains checks, and if any of them fail, the test will fail.
+  await transferRegionToRegionX(coretimeApi, regionXApi, alice, regionId);
 
-  const feeAssetItem = 0;
-  const weightLimit = 'Unlimited';
-  const reserveTransferToRegionX = coretimeApi.tx.polkadotXcm.limitedReserveTransferAssets(
-    { V3: { parents: 1, interior: { X1: { Parachain: 2000 } } } }, //dest
-    {
-      V3: {
-        parents: 0,
-        interior: {
-          X1: {
-            AccountId32: {
-              chain: 'Any',
-              id: receiverKeypair.pairs[0].publicKey,
-            },
-          },
-        },
-      },
-    }, //beneficiary
-    {
-      V3: [
-        {
-          id: {
-            Concrete: {
-              parents: 1,
-              interior: 'Here',
-            },
-          },
-          fun: {
-            Fungible: 10n ** 10n,
-          },
-        }, // ^^ fee payment asset
-        {
-          id: {
-            Concrete: {
-              parents: 0,
-              interior: { X1: { PalletInstance: 50 } },
-            },
-          },
-          fun: {
-            NonFungible: {
-              Index: getEncodedRegionId(regionId, coretimeApi).toString(),
-            },
-          },
-        },
-      ],
-    }, //asset
-    feeAssetItem,
-    weightLimit
-  );
-  await submitExtrinsic(alice, reserveTransferToRegionX, {});
-
-  await sleep(5000);
-
-  const requestRecord = regionXApi.tx.regions.requestRegionRecord(regionId);
-  await submitExtrinsic(alice, requestRecord, {});
-
-  let regions = await regionXApi.query.regions.regions.entries();
-  assert.equal(regions.length, 1);
-  assert.deepStrictEqual(regions[0][0].toHuman(), [regionId]);
-
-  let region = regions[0][1].toHuman() as any;
-  assert(region.owner == alice.address);
-  assert(typeof region.record.Pending === 'string');
-
-  // Check the data on the Coretime chain:
-  regions = await coretimeApi.query.broker.regions.entries();
-  assert.equal(regions.length, 1);
-  assert.deepStrictEqual(regions[0][0].toHuman(), [regionId]);
-  assert.equal((regions[0][1].toHuman() as any).owner, REGIONX_SOVEREIGN_ACCOUNT);
-
-  // Respond to the ISMP get request:
-  const request = await queryRequest(regionXApi, region.record.Pending);
-  await makeIsmpResponse(regionXApi, coretimeApi, request, alice.address);
-
-  // The record should be set after ISMP response:
-  regions = await regionXApi.query.regions.regions.entries();
-  region = regions[0][1].toHuman() as any;
-  assert(region.owner == alice.address);
-  assert.equal(region.record.Available.end, '66');
-  assert.equal(region.record.Available.paid, null);
-
-  // Transfer the region back to the Coretime chain:
-  const reserveTransferToCoretime = regionXApi.tx.polkadotXcm.limitedReserveTransferAssets(
-    { V3: { parents: 1, interior: { X1: { Parachain: 1005 } } } }, // dest
-    {
-      V3: {
-        parents: 0,
-        interior: {
-          X1: {
-            AccountId32: {
-              chain: 'Any',
-              id: receiverKeypair.pairs[0].publicKey,
-            },
-          },
-        },
-      },
-    }, // ^^ beneficiary
-    {
-      V3: [
-        {
-          id: {
-            Concrete: {
-              parents: 1,
-              interior: 'Here',
-            },
-          },
-          fun: {
-            Fungible: 10n ** 10n,
-          },
-        }, // ^^ fee payment asset
-        {
-          id: {
-            Concrete: {
-              parents: 1,
-              // chain: Rococo-Coretime, pallet: pallet_broker
-              interior: { X2: [{ Parachain: 1005 }, { PalletInstance: 50 }] },
-            },
-          },
-          fun: {
-            NonFungible: {
-              Index: getEncodedRegionId(regionId, regionXApi).toString(),
-            },
-          },
-        },
-      ],
-    }, // ^^ asset
-    feeAssetItem,
-    weightLimit
-  );
-  await submitExtrinsic(alice, reserveTransferToCoretime, {});
-  await sleep(5000);
-
-  regions = await regionXApi.query.regions.regions.entries();
-  assert.equal(regions.length, 0);
-
-  regions = await coretimeApi.query.broker.regions.entries();
-  assert.equal(regions.length, 1);
-  assert.deepStrictEqual(regions[0][0].toHuman(), [regionId]);
-  assert.equal((regions[0][1].toHuman() as any).owner, alice.address);
+  // Transferring back to the Coretime chain should work:
+  // NOTE: the function contains checks, and if any of them fail, the test will fail.
+  await transferRegionToCoretimeChain(coretimeApi, regionXApi, alice, regionId);
 }
 
 export { run };
